Rough start
This wouldn't pass any tests, if it had them, but it's a close enough start to make clear what I'm trying to do. There's some work to do to decide whether I can just pipe a load of secrets into an encrypted queue - for example, any admins could grab the secrets out of there and start messing with them.

On the other hand, storing them in a database enables the same thing, and this approach is needed to actually do the work...so I guess everyone involved would have to know what might happen. Hum.